Messed up the file reading a little bit.  Need to Fix Nexus and EMD formats.
